requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
PyMuPDF>=1.23.0

# Data validation and processing
pydantic>=2.0.0

# Excel export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Logging and utilities
colorama>=0.4.6

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
